Make external_id a required var in webhook-datasource module (#26)
Change summary:
----------------
The external_id variable was made optional so as to not break
onboarding backend. In the backend, we always ensure passing
the external_id to the module, because if not provided, the
validation for the same would fail via WIF.

Making this required now (as it should be) since backend is all merged.